+++ CHANGELOG.md
@@ -9,6 +9,10 @@ For more details on the contents of a release, see [the GitHub release page] (ht
 
 _**Note:** Yet to be released breaking changes appear here._
 
+**Breaking Changes**:
+- the `AbstractGraph.fit` method moved to `FitPlugin`, as well as the `minFitScale` and `maxFitScale` properties.
+The method now accepts a single parameter, mainly to minimize the need to pass many default values.
+
 ## 0.20.0
 
 Release date: `2025-05-16`
