+++ packages/core/src/view/AbstractGraph.ts
@@ -355,18 +355,6 @@ export abstract class AbstractGraph extends EventSource {
   multigraph = true;
 
   /**
-   * Specifies the minimum scale to be applied in {@link fit}. Set this to `null` to allow any value.
-   * @default 0.1
-   */
-  minFitScale: number | null = 0.1;
-
-  /**
-   * Specifies the maximum scale to be applied in {@link fit}. Set this to `null` to allow any value.
-   * @default 8
-   */
-  maxFitScale: number | null = 8;
-
-  /**
    * Specifies the {@link Image} for the image to be used to display a warning
    * overlay. See {@link setCellWarning}. Default value is Client.imageBasePath +
    * '/warning'.  The extension for the image depends on the platform. It is
@@ -799,134 +787,6 @@ export abstract class AbstractGraph extends EventSource {
       hCount * page.width + 2 + tr.x,
       vCount * page.height + 2 + tr.y
     );
-  }
-
-  /**
-   * Scales the graph such that the complete diagram fits into {@link AbstractGraph.container} and returns the current scale in the view.
-   * To fit an initial graph prior to rendering, set {@link GraphView.rendering} to `false` prior to changing the model
-   * and execute the following after changing the model.
-   *
-   * ```javascript
-   * graph.view.rendering = false;
-   * // here, change the model
-   * graph.fit();
-   * graph.view.rendering = true;
-   * graph.refresh();
-   * ```
-   *
-   * To fit and center the graph, use {@link FitPlugin.fitCenter}.
-   *
-   * @param border Optional number that specifies the border. Default is {@link border}.
-   * @param keepOrigin Optional boolean that specifies if the translate should be changed. Default is `false`.
-   * @param margin Optional margin in pixels. Default is `0`.
-   * @param enabled Optional boolean that specifies if the scale should be set or just returned. Default is `true`.
-   * @param ignoreWidth Optional boolean that specifies if the width should be ignored. Default is `false`.
-   * @param ignoreHeight Optional boolean that specifies if the height should be ignored. Default is `false`.
-   * @param maxHeight Optional maximum height.
-   */
-  fit(
-    border: number = this.getBorder(),
-    keepOrigin = false,
-    margin = 0,
-    enabled = true,
-    ignoreWidth = false,
-    ignoreHeight = false,
-    maxHeight: number | null = null
-  ): number {
-    const { container, view } = this;
-    if (container) {
-      // Adds spacing and border from css
-      const cssBorder = this.getBorderSizes();
-      let w1: number = container.offsetWidth - cssBorder.x - cssBorder.width - 1;
-      let h1: number =
-        maxHeight != null
-          ? maxHeight
-          : container.offsetHeight - cssBorder.y - cssBorder.height - 1;
-      let bounds = view.getGraphBounds();
-
-      if (bounds.width > 0 && bounds.height > 0) {
-        if (keepOrigin && bounds.x != null && bounds.y != null) {
-          bounds = bounds.clone();
-          bounds.width += bounds.x;
-          bounds.height += bounds.y;
-          bounds.x = 0;
-          bounds.y = 0;
-        }
-
-        // LATER: Use unscaled bounding boxes to fix rounding errors
-        const originalScale = view.scale;
-        let w2 = bounds.width / originalScale;
-        let h2 = bounds.height / originalScale;
-
-        // Fits to the size of the background image if required
-        if (this.backgroundImage) {
-          w2 = Math.max(w2, this.backgroundImage.width - bounds.x / originalScale);
-          h2 = Math.max(h2, this.backgroundImage.height - bounds.y / originalScale);
-        }
-
-        const b: number = (keepOrigin ? border : 2 * border) + margin + 1;
-
-        w1 -= b;
-        h1 -= b;
-
-        let newScale = ignoreWidth
-          ? h1 / h2
-          : ignoreHeight
-            ? w1 / w2
-            : Math.min(w1 / w2, h1 / h2);
-
-        if (this.minFitScale != null) {
-          newScale = Math.max(newScale, this.minFitScale);
-        }
-
-        if (this.maxFitScale != null) {
-          newScale = Math.min(newScale, this.maxFitScale);
-        }
-
-        if (enabled) {
-          if (!keepOrigin) {
-            if (!hasScrollbars(container)) {
-              const x0 =
-                bounds.x != null
-                  ? Math.floor(
-                      view.translate.x -
-                        bounds.x / originalScale +
-                        border / newScale +
-                        margin / 2
-                    )
-                  : border;
-              const y0 =
-                bounds.y != null
-                  ? Math.floor(
-                      view.translate.y -
-                        bounds.y / originalScale +
-                        border / newScale +
-                        margin / 2
-                    )
-                  : border;
-
-              view.scaleAndTranslate(newScale, x0, y0);
-            } else {
-              view.setScale(newScale);
-              const newBounds = this.getGraphBounds();
-
-              if (newBounds.x != null) {
-                container.scrollLeft = newBounds.x;
-              }
-
-              if (newBounds.y != null) {
-                container.scrollTop = newBounds.y;
-              }
-            }
-          } else if (view.scale != newScale) {
-            view.setScale(newScale);
-          }
-        } else {
-          return newScale;
-        }
-      }
-    }
-    return view.scale;
   }
 
   /**
